# backend/src/main/java/com/resumego/resume/dto/ResumeDTO.java
package com.resumego.resume.dto;

import java.time.LocalDateTime;

public record ResumeDTO(
        Long id,
        String title,
        Long targetJobDescriptionId,
        ResumeVersionDTO currentVersion,
        LocalDateTime createdAt,
        LocalDateTime updatedAt
) {
}