# harness-docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: 职达 Sprint 1—3 API
  version: 0.3.0-draft
  description: |
    职达 Sprint 1—3 核心闭环 API 评审草案。

    本文档由 AI 辅助整理，必须由三名成员人工审阅和修改后才能标记为 Accepted。
    当前阶段使用固定演示用户，不提供注册登录和文件上传。
servers:
  - url: http://localhost:8080
    description: 本地开发环境
...
paths:
...
  /api/v1/evidences:
    post:
      tags: [Evidence]
      summary: 创建能力证据
      operationId: createEvidence
      parameters:
        - $ref: "#/components/parameters/RequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEvidenceRequest"
      responses:
        "201":
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvidenceResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
    get:
      tags: [Evidence]
      summary: 查询演示用户的能力证据
      operationId: listEvidences
      parameters:
        - $ref: "#/components/parameters/RequestId"
      responses:
        "200":
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvidenceListResponse"
  /api/v1/evidences/{evidenceId}:
    get:
      tags: [Evidence]
      summary: 查询一条能力证据
      operationId: getEvidence
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/EvidenceId"
      responses:
        "200":
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvidenceResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Evidence]
      summary: 修改能力证据
      operationId: updateEvidence
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/EvidenceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEvidenceRequest"
      responses:
        "200":
          description: 修改成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvidenceResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/resumes:
    post:
      tags: [Resume]
      summary: 创建简历及首个版本
      operationId: createResume
      parameters:
        - $ref: "#/components/parameters/RequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateResumeRequest"
      responses:
        "201":
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumeResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/resumes/{resumeId}:
    get:
      tags: [Resume]
      summary: 查询简历及当前版本
      operationId: getResume
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/ResumeId"
      responses:
        "200":
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumeResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/resumes/{resumeId}/versions:
    get:
      tags: [Resume]
      summary: 查询简历版本列表
      operationId: listResumeVersions
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/ResumeId"
      responses:
        "200":
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumeVersionListResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/resume-versions/{versionId}:
    get:
      tags: [Resume]
      summary: 查询不可变简历版本
      operationId: getResumeVersion
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/VersionId"
      responses:
        "200":
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumeVersionResponse"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/job-matches/{matchId}/suggestions:
    post:
      tags: [Optimization]
      summary: 使用千问 Max 生成一次定向修改建议
      description: |
        建议必须引用能力证据。没有证据时只能返回 evidence_required，不得编造经历或数字。
      operationId: generateOptimizationSuggestions
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/MatchId"
      responses:
        "201":
          description: 建议生成成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuggestionListResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/AiOutputInvalid"
        "503":
          $ref: "#/components/responses/AiUnavailable"
...
components:
  parameters:
    RequestId:
      name: X-Request-Id
      in: header
      required: false
      description: 客户端可传入的请求关联 ID；缺失时由服务端生成
      schema:
        type: string
        maxLength: 64
...
    EvidenceId:
      name: evidenceId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Id"
    ResumeId:
      name: resumeId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Id"
    VersionId:
      name: versionId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Id"
...
    MatchId:
      name: matchId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Id"
...
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: 资源不存在或不属于当前演示用户
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    AiOutputInvalid:
      description: AI 返回内容未通过结构化校验
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    AiUnavailable:
      description: 模型超时、限流或供应商暂时不可用
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    Id:
      type: integer
      format: int64
      minimum: 1
    Timestamp:
      type: string
      format: date-time
    EvidenceType:
      type: string
      enum: [project, internship, competition, skill, other]
    SuggestionStatus:
      type: string
      enum: [pending, accepted, rejected, evidence_required]
...
    ApiMeta:
      type: object
      additionalProperties: false
      required: [requestId]
      properties:
        requestId:
          type: string
    CreateEvidenceRequest:
      type: object
      additionalProperties: false
      required: [type, title, action, skillTags]
      properties:
        type:
          $ref: "#/components/schemas/EvidenceType"
        title:
          type: string
          minLength: 1
          maxLength: 200
        situation:
          type: [string, "null"]
          maxLength: 5000
        action:
          type: string
          minLength: 1
          maxLength: 10000
        result:
          type: [string, "null"]
          maxLength: 5000
        skillTags:
          type: array
          maxItems: 30
          uniqueItems: true
          items:
            type: string
            minLength: 1
            maxLength: 50
        sourceNote:
          type: [string, "null"]
          maxLength: 500
    UpdateEvidenceRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        type:
          $ref: "#/components/schemas/EvidenceType"
        title:
          type: string
          minLength: 1
          maxLength: 200
        situation:
          type: [string, "null"]
          maxLength: 5000
        action:
          type: string
          minLength: 1
          maxLength: 10000
        result:
          type: [string, "null"]
          maxLength: 5000
        skillTags:
          type: array
          maxItems: 30
          uniqueItems: true
          items:
            type: string
            minLength: 1
            maxLength: 50
        sourceNote:
          type: [string, "null"]
          maxLength: 500
    Evidence:
      allOf:
        - $ref: "#/components/schemas/CreateEvidenceRequest"
        - type: object
          required: [id, createdAt, updatedAt]
          properties:
            id:
              $ref: "#/components/schemas/Id"
            createdAt:
              $ref: "#/components/schemas/Timestamp"
            updatedAt:
              $ref: "#/components/schemas/Timestamp"
    PersonalInfo:
      type: object
      additionalProperties: false
      required: [fullName]
      properties:
        fullName:
          type: string
          minLength: 1
          maxLength: 100
        email:
          type: [string, "null"]
          format: email
        phone:
          type: [string, "null"]
          maxLength: 30
        city:
          type: [string, "null"]
          maxLength: 100
    EducationItem:
      type: object
      additionalProperties: false
      required: [school, major]
      properties:
        school:
          type: string
          minLength: 1
          maxLength: 200
        major:
          type: string
          minLength: 1
          maxLength: 200
        degree:
          type: [string, "null"]
          maxLength: 100
        startDate:
          type: [string, "null"]
          format: date
        endDate:
          type: [string, "null"]
          format: date
    ResumeExperience:
      type: object
      additionalProperties: false
      required: [sectionKey, evidenceId, title, bullets]
      properties:
        sectionKey:
          type: string
          minLength: 1
          maxLength: 100
        evidenceId:
          $ref: "#/components/schemas/Id"
        title:
          type: string
          minLength: 1
          maxLength: 200
        bullets:
          type: array
          minItems: 1
          maxItems: 10
          items:
            type: string
            minLength: 1
            maxLength: 1000
    ResumeContent:
      type: object
      additionalProperties: false
      required: [personalInfo, education, experiences, skills]
      properties:
        personalInfo:
          $ref: "#/components/schemas/PersonalInfo"
        education:
          type: array
          maxItems: 10
          items:
            $ref: "#/components/schemas/EducationItem"
        experiences:
          type: array
          minItems: 1
          maxItems: 30
          items:
            $ref: "#/components/schemas/ResumeExperience"
        skills:
          type: array
          maxItems: 100
          uniqueItems: true
          items:
            type: string
            minLength: 1
            maxLength: 100
    CreateResumeRequest:
      type: object
      additionalProperties: false
      required: [title, content]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        content:
          $ref: "#/components/schemas/ResumeContent"
    ResumeVersion:
      type: object
      additionalProperties: false
      required:
        [id, resumeId, versionNo, content, createdByType, createdAt]
      properties:
        id:
          $ref: "#/components/schemas/Id"
        resumeId:
          $ref: "#/components/schemas/Id"
        parentVersionId:
          oneOf:
            - $ref: "#/components/schemas/Id"
            - type: "null"
        versionNo:
          type: integer
          minimum: 1
        content:
          $ref: "#/components/schemas/ResumeContent"
        changeSummary:
          type: [string, "null"]
          maxLength: 500
        createdByType:
          type: string
          enum: [user, ai_suggestion, import]
        createdAt:
          $ref: "#/components/schemas/Timestamp"
    Resume:
      type: object
      additionalProperties: false
      required: [id, title, currentVersion, createdAt, updatedAt]
      properties:
        id:
          $ref: "#/components/schemas/Id"
        title:
          type: string
        currentVersion:
          $ref: "#/components/schemas/ResumeVersion"
        createdAt:
          $ref: "#/components/schemas/Timestamp"
        updatedAt:
          $ref: "#/components/schemas/Timestamp"
...
    OptimizationSuggestion:
      type: object
      additionalProperties: false
      required:
        - id
        - matchId
        - resumeVersionId
        - sectionKey
        - originalText
        - reason
        - targetRequirement
        - status
        - promptVersion
      properties:
        id:
          $ref: "#/components/schemas/Id"
        matchId:
          $ref: "#/components/schemas/Id"
        resumeVersionId:
          $ref: "#/components/schemas/Id"
        evidenceId:
          oneOf:
            - $ref: "#/components/schemas/Id"
            - type: "null"
        sectionKey:
          type: string
        originalText:
          type: string
        suggestedText:
          type: [string, "null"]
        reason:
          type: string
        targetRequirement:
          type: string
        status:
          $ref: "#/components/schemas/SuggestionStatus"
        promptVersion:
          type: string
        acceptedVersionId:
          oneOf:
            - $ref: "#/components/schemas/Id"
            - type: "null"
    Error:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
          description: 稳定的机器可读错误码
        message:
          type: string
        fieldErrors:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [field, message]
            properties:
              field:
                type: string
              message:
                type: string
    EvidenceResponse:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/Evidence"
        meta:
          $ref: "#/components/schemas/ApiMeta"
    EvidenceListResponse:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Evidence"
        meta:
          $ref: "#/components/schemas/ApiMeta"
    ResumeResponse:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/Resume"
        meta:
          $ref: "#/components/schemas/ApiMeta"
    ResumeVersionResponse:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/ResumeVersion"
        meta:
          $ref: "#/components/schemas/ApiMeta"
    ResumeVersionListResponse:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ResumeVersion"
        meta:
          $ref: "#/components/schemas/ApiMeta"
...
    SuggestionListResponse:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/OptimizationSuggestion"
        meta:
          $ref: "#/components/schemas/ApiMeta"
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error, meta]
      properties:
        error:
          $ref: "#/components/schemas/Error"
        meta:
          $ref: "#/components/schemas/ApiMeta"